migration students and groups tables
