--- alembic/versions/2024_05_15_2244-af89fb135cd7_update_students_table_name.py
@@ -1,47 +1,0 @@
-"""Update students table name
-
-Revision ID: af89fb135cd7
-Revises: 50b03881dff6
-Create Date: 2024-05-15 22:44:09.553198
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = "af89fb135cd7"
-down_revision: Union[str, None] = "50b03881dff6"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "students",
-        sa.Column("first_name", sa.String(), nullable=False),
-        sa.Column("last_name", sa.String(), nullable=False),
-        sa.Column("group_id", sa.Integer(), nullable=True),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.drop_table("studentmodels")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "studentmodels",
-        sa.Column("first_name", sa.VARCHAR(), nullable=False),
-        sa.Column("last_name", sa.VARCHAR(), nullable=False),
-        sa.Column("group_id", sa.INTEGER(), nullable=True),
-        sa.Column("id", sa.INTEGER(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.drop_table("students")
-    # ### end Alembic commands ###
