+++ alembic/versions/2024_05_15_2257-c3db6a5c0fb3_create_students_and_groups_tables.py
@@ -1,0 +1,42 @@
+"""Create students and groups tables
+
+Revision ID: c3db6a5c0fb3
+Revises: 
+Create Date: 2024-05-15 22:57:18.578060
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = "c3db6a5c0fb3"
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "students",
+        sa.Column("first_name", sa.String(length=20), nullable=False),
+        sa.Column("last_name", sa.String(length=20), nullable=False),
+        sa.Column("group_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["group_id"],
+            ["groups.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("students")
+    # ### end Alembic commands ###
